requirements.txt
Django==3.2.4
django-cors-headers==3.7.0 
gunicorn==20.1.0
whitenoise==5.2.0
psycopg2==2.8.6
django-heroku==0.3.1
psycopg2-binary==2.9
dj-database-url==0.5.0
python-dotenv==0.17.1
djangorestframework==3.12.4